% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bugs.R
\name{densityplot}
\alias{densityplot}
\title{Density plots for the posteriors in a \code{bugs} object}
\usage{
densityplot(x, parameter = NULL, ...)
}
\arguments{
\item{x}{an object of class `bugs', see \code{\link{bugs}} for details}

\item{parameter}{a string with the name of the parameter for which to show
the density plot}

\item{...}{further arguments to \code{\link{densityplot}}}
}
\description{
Density plots for the posteriors in a \code{bugs} object
}
\seealso{
\code{\link{bugs}}
}
