% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bugs.R
\name{posteriorplot}
\alias{posteriorplot}
\title{Various plots for the posteriors in a \code{bugs} object}
\usage{
posteriorplot(x, parameter = NULL, plot = "density", ...)
}
\arguments{
\item{x}{an object of class `bugs', see \code{\link{bugs}} for details}

\item{parameter}{a string with the name of the parameter for which to show
the density plot}

\item{plot}{the type of plot (options are 'density' (default) or 
'hist' for a binned barplot of the posterior)}

\item{...}{further arguments to \code{\link{densityplot}}}
}
\description{
Various plots for the posteriors in a \code{bugs} object
}
\seealso{
\code{\link{bugs}}
}
\author{
Gianluca Baio
}
