% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bugs.R
\name{traceplot}
\alias{traceplot}
\title{Tidyverse based function to do traceplots}
\usage{
traceplot(x, parameter = NULL, ...)
}
\arguments{
\item{x}{an object of class `bugs', see \code{\link{bugs}} for details}

\item{parameter}{a string with the name of the parameter for which to show
the traceplot}

\item{...}{further arguments to \code{\link{traceplot}}}
}
\description{
Traceplot for a \code{bugs} object
}
\seealso{
\code{\link{bugs}}
}
\author{
Gianluca Baio
}
